requirements.txt
requests
python-dotenv
supabase
beautifulsoup4